Add files via upload
RTC Clock and more helpfull things. In the cas you like to use the Badger2040 as an Door or Desk Plate/Sign you can set the text in using the inttext.py in combination with the textonbadger.py. Run textonbadger.py an you will get ask for the Text to print on your Badger2040, like a Room Numer, Employee Name an the Job Title. textonbadger_led.py let blink the Badger2040 LED bevor input Text.